backend/requirements.txt
aioredis==1.3.1
amqp==5.1.1
anyio==3.7.1
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
awaits==0.0.1
beautifulsoup4==4.6.3
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==3.2.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coloredlogs==15.0.1
colorlog==4.0.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
croniter==1.3.5
cryptography==38.0.4
curlify==2.2.1
daphne==3.0.2
ddddocr==1.4.10
Django==3.2.1
django-bulk-update==2.2.0
django-celery-beat==2.4.0
django-celery-results==2.2.0
django-cors-headers==4.0.0
django-jsonfield==1.4.0
django-log-request-id==2.0.0
django-model-utils==4.0.0
django-simpleui==2020.7
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.5
exceptiongroup==1.1.3
Faker==7.0.1
Flask==2.2.5
flatbuffers==23.5.26
flower==1.2.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hiredis==2.2.3
hpack==4.0.0
httptools==0.6.0
humanfriendly==10.0
humanize==4.6.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
json5==0.9.5
jsonpath==0.82
kaitaistruct==0.10
kombu==5.2.4
ldap3==2.9.1
loguru==0.5.1
MarkupSafe==2.1.1
mitmproxy==9.0.1
mitmproxy_wireguard==0.1.23
mpmath==1.3.0
msgpack==1.0.5
mysqlclient==2.1.1
numpy==1.26.2
onnxruntime==1.16.3
opencv-python-headless==4.8.1.78
packaging==23.0
passlib==1.7.4
Pillow==10.1.0
pipdeptree==2.11.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.24.2
publicsuffix2==2.20191221
py-mini-racer==0.4.0
py-mock==1.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.9.0
pydash==5.0.1
PyJWT==1.7.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyPEG2==2.15.2
pyperclip==1.8.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7
PyYAML==5.3.1
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.28.2
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sentry-sdk==1.28.1
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
sympy==1.12
text-unidecode==1.3
toml==0.10.2
tornado==6.3.2
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.14
urwid==2.1.2
user-agents==2.2.0
uvicorn==0.17.6
uvloop==0.17.0
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==11.0.3
Werkzeug==2.3.7
wsproto==1.2.0
xmltodict==0.12.0
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
zstandard==0.19.0
